Add tox and supporting files
This adds tox.ini with py27 and pep8 environments. It also adds
the files necessary for supporting tox, and for allowing to test
without mysql things installed.

Not sure how setup.py was previously ignored, or how it is/was
generated by distutils; it's commonly *used* by distutils to
install things. Regardless of how it is created, it shouldn't
be ignored.

Caveats:
* The setup.py I added is pretty arbitrary; I'm open to suggestions
  here.
* I feel like removing mysqldb from requirements.txt is going to
  docs work, but unfortunately the docs are not in this repo. :(
* Would love test_settings.py to import everything from settings
  and override as needed; not done yet.