Return a useful error message if the user has 2FA enabled on RED
RED recently introduced 2FA, but of course WM2 can't handle this (yet... although it would be a little complicated, I suspect).

I added a check in CustomWhatAPI._login() for the 2FA form HTML in the response and returned a relevant message in the exception.

I also added a generic message if the login fails for any reason. Since this message propogates all the way up to the front page / Log page, an empty error isn't very helpful.